Make_theta, Make_gamma
Added make_theta and make_gamma function in the general_load_bucket.py, the gamma is sampled according to the histogram of input gamma, theta is generated randomly as load_bucket.py but with different Ns in each bucket.
To illustrate the sampled gamma value follows the distribution of original gamma value, here I input an unrealistic normal gamma distribution in general_load_bucket instead of the gamma of the input particle file. I plotted the input and sampled gamma distribution, and they should obey the same distribution.